+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+- Update `parse-dockerfile@latest` to 0.1.2.
+
 - Update `parse-changelog@latest` to 0.6.13.
 
 - Update `cargo-no-dev-deps@latest` to 0.2.17.