+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+- Update `cargo-no-dev-deps@latest` to 0.2.17.
+
 - Update `cargo-minimal-versions@latest` to 0.1.32.
 
 - Update `cargo-llvm-cov@latest` to 0.6.19.