+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+## [2.60.0] - 2025-09-08
+
 - Add `fallback: cargo-install` input option to use `cargo install` instead of `cargo-binstall` for fallback.
 
 - Support powerpc64le/riscv64/s390x Linux runners. The following tools are now supported, and for other tools the `cargo install` fallback mentioned above is used for now.
@@ -4441,7 +4443,8 @@ Note: This release is considered a breaking change because installing on version
 
 Initial release
 
-[Unreleased]: https://github.com/taiki-e/install-action/compare/v2.59.1...HEAD
+[Unreleased]: https://github.com/taiki-e/install-action/compare/v2.60.0...HEAD
+[2.60.0]: https://github.com/taiki-e/install-action/compare/v2.59.1...v2.60.0
 [2.59.1]: https://github.com/taiki-e/install-action/compare/v2.59.0...v2.59.1
 [2.59.0]: https://github.com/taiki-e/install-action/compare/v2.58.33...v2.59.0
 [2.58.33]: https://github.com/taiki-e/install-action/compare/v2.58.32...v2.58.33